Added URI shortcuts
Two types of shortcuts, normal and native.

Normal shortcuts allow users to enter '@google' to quickly go to 'google.com'. They are similar to bookmarks.

Native shortcuts operate similarly but are designed to be the primary address of the native page. This allows the SettingsPage to be represented with '://settings' instead of ':://GameBarBrowser.Settings.SettingsPage'. Thinking about it, attributes would have probably been more ideal.