Added a scroll bar to scroll through tabs
This is one fix to prevent tabs from being unreachable.

There is an issue where the scroll bar doesn't increase its padding when creating a new tab. The padding is increased once a second tab is created or the window is shrinked. This issue is fairly odd as the padding for the New Tab button does get increased.