History complete
History has been added.

Users have the options for history to be recorded, and if it should allow duplicated entries.

A webpage is added to the history after it loads.

History can be viewed in the Library (it's intended for it to be ordered in the time the page was last visited).

History items can be removed individually or by a TimeSpan (last day, last week, etc).

When creating the History feature, I noticed a lot of work could be done refactoring Library to reduce messiness and redundancy. I may focus on this later if I have available time.