Pages seem to navigate properly
Pages between the Native and Web Views now seem to navigate correctly.
Except it all breaks when the timeline has been changed.

I give up. I'm going to work on a totally different solution where I store the URLs of every site I visit and manage the entire timeline my self. This should be less performant as the browser will always create a new page, instead of using a cached page, but it's the only solution I can think of.