Refresh Library, removed refresh button for Native pages
Added a refresh button for the Library, which reloads Bookmarks and History.

Disabled the tab refresh button for Native pages as they don't have the ability to be reloaded. (Maybe if I made my own timeline it would, but after investing so much time in the current navigation solution, cba)